Fix error editing files in place in other directories
When editing in place, it creates a temporary file named the same as the
nbt file with .txt appended. It would try to create this directly from
the path passed to it, which works if editing files in the same
directory ('nbted somefile.nbt') but not if the file is in another
directory ('nbted somedir/somefile.nbt',) causing it to give an error
and exit before giving the user a chance to edit the file.

This simply changes it so that it now only creates the temporary file
with the filename part of its path.