Rewrite the tokenizer
This removes the dependency on regex. The old and simple regex tokenizer
worked fine but had the major problem that I could not figure out a way
to count lines encountered using it. As I want error messages to provide
line numbers, I've written the tokenizer from scratch to one that
manually parses the bytes in the text file one by one.

Also added a bunch of new tests to test the new tokenizer.

Barring bugs, the only behavioral change should be that nbted now
supports bare strings. But as bare strings may be disallowed in the
future, they should not be relied on.